Add exception handler

+++ src/main/java/ua/bondoleks/loyaltySystem/service/PurchaseService.java
@@ -2,9 +2,8 @@ package ua.bondoleks.loyaltySystem.service;
 
 import org.springframework.stereotype.Service;
 import ua.bondoleks.loyaltySystem.entity.Purchase;
-import ua.bondoleks.loyaltySystem.entity.LUser;
+import ua.bondoleks.loyaltySystem.exception.PurchaseNotFoundException;
 import ua.bondoleks.loyaltySystem.repository.PurchaseRepository;
-import ua.bondoleks.loyaltySystem.repository.UserRepository;
 
 import java.util.List;
 
@@ -12,11 +11,9 @@ import java.util.List;
 public class PurchaseService {
 
     private final PurchaseRepository purchaseRepository;
-    private final UserRepository userRepository;
 
-    public PurchaseService(PurchaseRepository purchaseRepository, UserRepository userRepository) {
+    public PurchaseService(PurchaseRepository purchaseRepository) {
         this.purchaseRepository = purchaseRepository;
-        this.userRepository = userRepository;
     }
 
     public List<Purchase> getAllPurchases() {
@@ -24,16 +21,11 @@ public class PurchaseService {
     }
 
     public Purchase getPurchaseById(Long id) {
-        return purchaseRepository.findById(id).orElseThrow(() -> new RuntimeException("Purchase not found"));
+        return purchaseRepository.findById(id)
+                .orElseThrow(() -> new PurchaseNotFoundException("Purchase with ID " + id + " not found"));
     }
 
     public void createPurchase(Purchase purchase) {
-        String phoneNumber = purchase.getUser().getPhoneNumber();
-        LUser user = userRepository.findByPhoneNumber(phoneNumber).orElseGet(() -> {
-            LUser newUser = purchase.getUser();
-            return userRepository.save(newUser);
-        });
-        purchase.setUser(user);
         purchaseRepository.save(purchase);
     }
 
@@ -41,13 +33,13 @@ public class PurchaseService {
         Purchase purchase = getPurchaseById(id);
         purchase.setItems(purchaseDetails.getItems());
         purchase.setTotalAmount(purchaseDetails.getTotalAmount());
-        purchase.setBonusAmount(purchaseDetails.getBonusAmount());
-        purchase.setReturned(purchaseDetails.isReturned());
-        purchase.setStoreType(purchaseDetails.getStoreType());
         purchaseRepository.save(purchase);
     }
 
     public void deletePurchase(Long id) {
+        if (!purchaseRepository.existsById(id)) {
+            throw new PurchaseNotFoundException("Purchase with ID " + id + " not found");
+        }
         purchaseRepository.deleteById(id);
     }
 }
